# chore: performance optimisation for set_data action