PR Checklist

- [x] PR title descriptive (can be used in release notes)

## Description

Optimises `set_data` action logic, which should make it more performant when applying large numbers of updates at once.

## Git Issues

Closes #

## Screenshots/Videos

@ChrisMarsh82 should understand the following example:

On the `plh_facilitator_ph` deployment, when switching between parent groups on the homepage, the changes in this PR bring a significant performance gain – a drop in processing the changes from ~4 seconds to ~1 second (still quite slow)

| master | PR branch |
|---|---|
| <img width="741" alt="Screenshot 2025-05-28 at 19 07 37" src="https://github.com/user-attachments/assets/37499534-5cdc-43f1-b626-baa2b934156e" /> | <img width="738" alt="Screenshot 2025-05-28 at 19 06 20" src="https://github.com/user-attachments/assets/f90537b8-08e4-445b-919c-c4b3b503e8cb" /> |