@jfmcquade 2025-06-03
> > @ChrisMarsh82 should understand the following example:
> > On the `plh_facilitator_ph` deployment, when switching between parent groups on the homepage, the changes in this PR bring a significant performance gain – a drop in processing the changes from ~4 seconds to ~1 second (still quite slow)
> 
> Any way you can help me understand too please? 😅 I have the `plh_facilitator_ph` running locally but don't see any way to change between parent groups. Would be keen to test the performance improvements and see if I can make any further recommendations

OK, you're allowed...

Here's a screen recording showing what's required from a fresh install: setting your user role to "System Tester" and then adding at least 2 parent groups. After that, you can switch between parent groups via the dropdown at the top of the home screen:

https://github.com/user-attachments/assets/103cd09a-401a-4635-a09c-7daa732da96a

I achieved the logs that I've screenshotted in the PR description by bookending the default exported function from `src/app/shared/services/dynamic-data/actions/set_data.action.ts` like this:

<img width="800" alt="Screenshot 2025-06-03 at 15 21 43" src="https://github.com/user-attachments/assets/fe039a79-951d-43ac-b87c-a0c2018d7fcd" />