@chrismclarke 2025-06-05
> 
> OK, you're allowed...
> 
> Here's a screen recording showing what's required from a fresh install: setting your user role to "System Tester" and then adding at least 2 parent groups. After that, you can switch between parent groups via the dropdown at the top of the home screen. The logic triggered is authored on row `14` of this [menu_task_flow](https://docs.google.com/spreadsheets/d/1JiKjLEZucBCeFuErhDnIjfTB3dSCrrxg9h0-YHlA7ns/edit?gid=0#gid=0) template. Essentially, it's setting a value for a single property for each row in a 200+ row data list.
> 

Thanks @jfmcquade, always nice to be in with the in-crowd. Also wow, yeah that's one slow update! Even with the fix in this branch my laptop struggles so would be good to optimise a bit better for sure. A couple suggestions

1. Lets try to get https://github.com/IDEMSInternational/open-app-builder/pull/2948 merged to fix issue with console errors and some of the data-items refs that are causing issue

2. I'm guessing this issue was a bit motivation for asking for a `filter` operator on the `set_data` action. That would help to streamline quite a bit, but as that's currently held up a bit it might be worth adding an automated filter to the action so that it first calculates an optimistic update (compare server data with rows for processing), and filters out all ops that would be unchanged. I think can probably use that instead of the custom update combination methods

3. Use a bulk processing method on the db itself. I can see `bulkUpsert` already exists on the service, so I think likely could just use that instead of the custom batching (or exposing a similar `bulkUpdate` if we there is good reason to provide separate update/upsert bulk ops for this use case (I can't really think of one myself, given that the set action already targets specific ids so shouldn't be any risk of accidentally inserting unwanted data)

Let me know if you're happy to work on any of these optimisations as part of this PR, or if you would prefer I create issues for standalone dev